added check for inactive/ active sw on login attempt or on attempt to create new sw

+++ src/pages/LoginWithSkillWallet.tsx
@@ -11,8 +11,9 @@ import {
   setUserName,
   setUserProfilePicture,
   showDialog,
+  setTokenId,
 } from '../store/sw-auth.reducer';
-import { fetchSkillWallet } from '../services/web3/web3Service';
+import { checkForInactiveSkillWallet, fetchSkillWallet } from '../services/web3/web3Service';
 import { ReactComponent as MetaMaskIcon } from '../assets/metamask.svg';
 import { ReactComponent as PortisIcon } from '../assets/portis_icon.svg';
 
@@ -30,35 +31,43 @@ const LoginWithSkillWallet: React.FunctionComponent = (props) => {
 
   const handleMetamaskClick = async () => {
     dispatch(setLoading(true));
+
     const { ethereum } = window;
     try {
       await window.ethereum.request({ method: 'eth_requestAccounts' });
       if (ethereum.selectedAddress) {
-        await fetchSkillWallet(ethereum.selectedAddress)
-          .then((result) => {
-            dispatch(setSkillWallet(result));
-            dispatch(setUserName(result.nickname));
-            dispatch(setUserProfilePicture(result.imageUrl));
-            dispatch(setLoggedIn(true));
-            dispatch(showDialog(false));
-            window.sessionStorage.setItem('skillWallet', JSON.stringify(result));
-            console.log(result);
-            history.push('/');
-            const event = new CustomEvent('onSkillwalletLogin', {
-              composed: true,
-              cancelable: true,
-              bubbles: true,
-              detail: true,
+        const res = await checkForInactiveSkillWallet(ethereum.selectedAddress);
+        if (res && res.inactiveSkillWalletExists) {
+          dispatch(setTokenId(res.tokenId.toString()));
+          history.push('/qr');
+        } else {
+          await fetchSkillWallet(ethereum.selectedAddress)
+            .then((result) => {
+              dispatch(setSkillWallet(result));
+              dispatch(setUserName(result.nickname));
+              dispatch(setUserProfilePicture(result.imageUrl));
+              dispatch(setLoggedIn(true));
+              dispatch(showDialog(false));
+              window.sessionStorage.setItem('skillWallet', JSON.stringify(result));
+              console.log(result);
+              history.push('/');
+              const event = new CustomEvent('onSkillwalletLogin', {
+                composed: true,
+                cancelable: true,
+                bubbles: true,
+                detail: true,
+              });
+              console.log('sending login event');
+              window.dispatchEvent(event);
+              dispatch(setLoading(false));
+            })
+            .catch((e) => {
+              setErrorData({ message: 'Failed to retrieve SkillWallet' });
+            })
+            .finally(() => {
+              dispatch(setLoading(false));
             });
-            console.log('sending login event');
-            window.dispatchEvent(event);
-          })
-          .catch(() => {
-            setErrorData({ message: 'Failed to retrieve SkillWallet' });
-          })
-          .finally(() => {
-            dispatch(setLoading(false));
-          });
+        }
       }
       dispatch(setLoading(false));
     } catch (error) {
